Fixes for voyager_test issues
Now gives similar results, see #51. Issue was
1) drift rate calculation didn't work for negative frequency steps.
2) DC bin needs to be blanked. This is now done with sk_flag

+++ hyperseti/normalize.py
@@ -50,7 +50,7 @@ def normalize(data,  mask=None, padding=0):
     logger.debug(f"flag fraction correction factor: {flag_correction}")
     
     #  Apply to original data
-    d_gpu = (d_gpu - d_mean) / d_std
+    d_gpu = ((d_gpu - d_mean) / d_std).astype('float32')
     t1 = time.time()
     logger.info(f"Normalisation time: {(t1-t0)*1e3:2.2f}ms")
